domain/src/main/kotlin/com/seanshubin/kotlin/tryme/domain/cycles/CycleUtil.kt
package com.seanshubin.kotlin.tryme.domain.cycles

object CycleUtil {
    fun <T : Comparable<T>> findCycles(edges: Set<Pair<T, T>>): List<List<T>> {
        val vertices: List<T> = edges.flatMap { listOf(it.first, it.second) }.sorted().distinct()
        val adjacencyMatrix: MutableList<MutableList<Boolean>> = vertices.map { row ->
            vertices.map { column ->
                edges.contains(row to column)
            }.toMutableList()
        }.toMutableList()
        vertices.indices.forEach { k ->
            vertices.indices.forEach { i ->
                vertices.indices.forEach { j ->
                    if (adjacencyMatrix[i][k] && adjacencyMatrix[k][j]) {
                        adjacencyMatrix[i][j] = true
                    }
                }
            }
        }
        val inCycle = vertices.indices.filter { adjacencyMatrix[it][it] }
        val cycles = inCycle.map { target ->
            val thisCanReachOther = vertices.indices.filter { adjacencyMatrix[target][it] }.toSet()
            val otherCanReachThis = vertices.indices.filter { adjacencyMatrix[it][target] }.toSet()
            val cycleIndices = thisCanReachOther intersect otherCanReachThis
            val cycle = cycleIndices.map { vertices[it] }
            cycle
        }
        return cycles.map { it.sorted() }.distinct()
    }
}
